requirements.txt
evdev==0.4.5
python-daemon==1.6.1
